@@ -45,4 +45,7 @@ class App extends React.Component {
   }
 }
 
-ReactDOM.render(<App />, document.root);
+const root = ReactDOM.createRoot(
+  document.getElementById('root')
+);
+root.render(<App />);
