@@ -7,23 +7,35 @@ import {useSandpack} from '@codesandbox/sandpack-react';
 import {IconArrowSmall} from '../../Icon/IconArrowSmall';
 export interface DownloadButtonProps {}
 
-let initialIsSupported = false;
+let supportsImportMap: boolean | void;
+
+function useSupportsImportMap() {
+  function subscribe() {
+    // It never updates.
+    return () => {};
+  }
+  function getCurrentValue() {
+    if (supportsImportMap === undefined) {
+      supportsImportMap =
+        (HTMLScriptElement as any).supports &&
+        (HTMLScriptElement as any).supports('importmap');
+    }
+    return supportsImportMap;
+  }
+  function getServerSnapshot() {
+    return false;
+  }
+
+  return React.useSyncExternalStore(
+    subscribe,
+    getCurrentValue,
+    getServerSnapshot
+  );
+}
 
 export const DownloadButton: React.FC<DownloadButtonProps> = () => {
   const {sandpack} = useSandpack();
-  const [supported, setSupported] = React.useState(initialIsSupported);
-  React.useEffect(() => {
-    // This detection will work in Chrome 97+
-    if (
-      !supported &&
-      (HTMLScriptElement as any).supports &&
-      (HTMLScriptElement as any).supports('importmap')
-    ) {
-      setSupported(true);
-      initialIsSupported = true;
-    }
-  }, [supported]);
-
+  const supported = useSupportsImportMap();
   if (!supported) {
     return null;
   }
