@@ -5,23 +5,24 @@
 import * as React from 'react';
 import {AppProps} from 'next/app';
 import {useRouter} from 'next/router';
-// @ts-ignore
-import galite from 'ga-lite';
+import {ga} from '../utils/analytics';
 import '@docsearch/css';
 import '../styles/algolia.css';
 import '../styles/index.css';
 import '../styles/sandpack.css';
 import '@codesandbox/sandpack-react/dist/index.css';
 
-const EmptyAppShell: React.FC = ({children}) => <>{children}</>;
+const EmptyAppShell = ({children}: {children: React.ReactNode}) => (
+  <>{children}</>
+);
 
 if (typeof window !== 'undefined') {
   if (process.env.NODE_ENV === 'production') {
-    galite('create', process.env.NEXT_PUBLIC_GA_TRACKING_ID, 'auto');
+    ga('create', process.env.NEXT_PUBLIC_GA_TRACKING_ID, 'auto');
   }
   const terminationEvent = 'onpagehide' in window ? 'pagehide' : 'unload';
   window.addEventListener(terminationEvent, function () {
-    galite('send', 'timing', 'JS Dependencies', 'unload');
+    ga('send', 'timing', 'JS Dependencies', 'unload');
   });
 }
 
@@ -29,8 +30,8 @@ export default function MyApp({Component, pageProps}: AppProps) {
   const router = useRouter();
   React.useEffect(() => {
     const handleRouteChange = (url: string) => {
-      galite('set', 'page', url);
-      galite('send', 'pageview');
+      ga('set', 'page', url);
+      ga('send', 'pageview');
     };
     router.events.on('routeChangeComplete', handleRouteChange);
     return () => {
