@@ -4,18 +4,18 @@
 
 import React from 'react';
 
-const addString = (list: React.ReactNodeArray, string: string) =>
+const addString = (list: React.ReactNode[], string: string) =>
   list.push(<span key={`${list.length}-${string}`}>{string}</span>);
 
 function toCommaSeparatedList<Item>(
   array: Item[],
-  renderCallback: (item: Item, index: number) => void
-): Array<any> {
+  renderCallback: (item: Item, index: number) => React.ReactNode
+): React.ReactNode[] {
   if (array.length <= 1) {
     return array.map(renderCallback);
   }
 
-  const list: any = [];
+  const list: React.ReactNode[] = [];
 
   array.forEach((item, index) => {
     if (index === array.length - 1) {
